#67 Split models into scoped dto files
