+++ src/quiz/quiz.service.ts
@@ -8,7 +8,7 @@ import {
   QuizImageType,
   QuizType,
 } from '@prisma/client';
-import { Quiz, QuizCompletion, QuizFilters, QuizImage } from '../models';
+import { Quiz, QuizCompletion, QuizFilters, QuizImage } from './quiz.dto';
 import { S3FileService } from '../file/s3.service';
 import { persistence } from '../persistence/persistence';
